random: move random bytes generation to bytes module
