+++ src/v/bytes/CMakeLists.txt
@@ -7,4 +7,15 @@ v_cc_library(
     absl::hash
     Seastar::seastar
   )
+
+v_cc_library(
+  NAME
+    bytes_random
+  SRCS
+    random.cc
+  DEPS
+    v::random
+    v::bytes
+)
+
 add_subdirectory(tests)
