+++ src/v/kafka/server/tests/group_metadata_serialization_test.cc
@@ -8,6 +8,7 @@
 // by the Apache License, Version 2.0
 
 #include "bytes/bytes.h"
+#include "bytes/random.h"
 #include "kafka/protocol/wire.h"
 #include "kafka/server/group_metadata.h"
 #include "kafka/server/server.h"
