+++ src/v/raft/tests/offset_translator_tests.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/random.h"
 #include "model/fundamental.h"
 #include "raft/offset_translator.h"
 #include "random/generators.h"
