+++ src/v/security/scram_authenticator.cc
@@ -11,6 +11,7 @@
 #include "security/scram_authenticator.h"
 
 #include "base/vlog.h"
+#include "random/generators.h"
 #include "security/credential_store.h"
 #include "security/errc.h"
 #include "security/logger.h"
