+++ src/v/storage/tests/batch_cache_test.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/random.h"
 #include "cluster/simple_batch_builder.h"
 #include "model/fundamental.h"
 #include "model/record.h"
