+++ src/v/storage/tests/compaction_idx_bench.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/random.h"
 #include "model/fundamental.h"
 #include "random/generators.h"
 #include "storage/compacted_index.h"
