+++ src/v/storage/tests/compaction_index_format_tests.cc
@@ -10,6 +10,7 @@
 #include "base/units.h"
 #include "bytes/bytes.h"
 #include "bytes/iobuf_parser.h"
+#include "bytes/random.h"
 #include "random/generators.h"
 #include "reflection/adl.h"
 #include "storage/compacted_index.h"
