+++ src/v/storage/tests/compaction_reducer_test.cc
@@ -1,3 +1,14 @@
+/*
+ * Copyright 2023 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+#include "bytes/random.h"
 #include "random/generators.h"
 #include "storage/compacted_index.h"
 #include "storage/compaction_reducers.h"
