+++ src/v/storage/tests/kvstore_test.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/random.h"
 #include "config/configuration.h"
 #include "random/generators.h"
 #include "reflection/adl.h"
