+++ src/v/utils/tests/base64_test.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/random.h"
 #include "random/generators.h"
 #include "utils/base64.h"
